+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 from local variables.
 - `__offsetof` for low level needs (works like `offsetof` in C).
 - vfmt now preserves empty lines, like gofmt.
+- Support for compile time environment variables via `$env('ENV_VAR')`. 
 
 ## V 0.2.1
 *30 Dec 2020*